Inference/Enemble: make it possible to use arbitrary amount of models
Change the API of EnsembleV1 to get a named list of models instead of
hardcoded TCVIS and NOTCVIS. During ensembling, a mean of all models
is calculated for each pixel.

The CLI tools parameters still keep the TCVIS/NOTCVIS model arguments,
but its now possible to leave one empty. In that case, there is essentially
no ensembling and the output is just the result of one model.

fixes #90