add the option to use a basic stdout logging handler
On albedo/Slurm the output is piped to a file, and the log width is
basically shrunk to about 80 chars which makes the Rich output
somewhat unreadable because the actual message has even less
space due to the column format of the rich ouput.

add --log-plain to the CLI inference tools, which will use
a boring stream handler for the output.