Add inference metadata to exports
export_tile() accepts a new parameter "metadata", a dictionary of
key/value-pairs which should be written to the export dataset. It will
write this to the export rasters GeoTIFFS as GDAL attributes.
Additionally, an export band "metadata" was added, this puts a metadata
JSON file alongside the other files in the output directory.

The metadatitself will be retrieved from the pipeline, a default set of
data is configured in BasePipeline._result_metadata(). Subclasses can
add to the metadata, as already done in
Sentinel2Pipeline._result_metadata().